Simplify dataset class
1. Remove other dataset classes and merge into one Dataset class. It
downloads the desired dataset if it's not already in the keras dataset
folder. The reason tf.keras.utils.get_file is not used for enwik9 is
because an error occurs when extracting the zip file on Windows,
dunno about other environments.

2. Added convert to ints method for the text dataset and convert to
tensor dataset, so it will be easier to utilize tf.data.Dataset methods
on the data.